pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "echodraft"
version = "0.1.0"
description = "EchoDraft — a feedback-aware, style-learning writing agent."
authors = [{name = "You"}]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "typer>=0.9.0",
  "python-dotenv>=1.0.1",
  # Core stack to add as you wire features (commented until used):
  # "langchain>=0.3.0",
  # "langgraph>=0.2.0",
  # "langsmith>=0.1.100",
  # "tiktoken>=0.7.0",
  # "numpy>=1.26.0",
  # "pydantic>=2.6.0",
  # "sentence-transformers>=3.0.0",
]

[project.scripts]
echodraft = "echodraft.ui.cli:app"
